Web_buzuluk/requirements.txt
Flask==3.0.2
fdb==2.0.2
requests==2.31.0
python-dateutil==2.8.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1 